refactor(communicationmapping): move wizards to CommunicationMapping.ts
